fix: Review fixes and revert to light theme
