+++ src/app/(components)/Header.tsx
@@ -2,17 +2,38 @@ import Link from "next/link";
 
 const Header = () => {
   return (
-    <div className="container text-center">
-      <div className="flex items-center">
-        <Link
-          href="/"
-          className="bg-sky-500 hover:bg-sky-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-        >
-          Back to home
-        </Link>
-        <h1 className="font-bold text-xl mx-auto">Email signature tool</h1>
-      </div>
-    </div>
+    <nav className="flex flex-wrap gap-3 justify-center py-4 bg-slate-800">
+      <Link
+        href="/signatures/wr"
+        className="bg-green-500 hover:bg-green-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
+      >
+        Wellreceived
+      </Link>
+      <Link
+        href="/signatures/sfca"
+        className="bg-black hover:bg-slate-900 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
+      >
+        Serviceforge CA
+      </Link>
+      <Link
+        href="#"
+        className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
+      >
+        Answerconnect US - WIP 🛠️
+      </Link>
+      <Link
+        href="#"
+        className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
+      >
+        Answerconnect UK - WIP 🛠️
+      </Link>
+      <Link
+        href="#"
+        className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
+      >
+        Answerforce US - WIP 🛠️
+      </Link>
+    </nav>
   );
 };
