+++ src/app/layout.tsx
@@ -18,38 +18,6 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <nav className="flex flex-wrap gap-3 justify-center py-4 bg-slate-800">
-          <Link
-            href="/signatures/wr"
-            className="bg-green-500 hover:bg-green-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-          >
-            Wellreceived
-          </Link>
-          <Link
-            href="/signatures/sfca"
-            className="bg-black hover:bg-slate-900 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-          >
-            Serviceforge CA
-          </Link>
-          <Link
-            href="#"
-            className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-          >
-            Answerconnect US - WIP 🛠️
-          </Link>
-          <Link
-            href="#"
-            className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-          >
-            Answerconnect UK - WIP 🛠️
-          </Link>
-          <Link
-            href="#"
-            className="bg-gray-500 hover:bg-gray-700 px-5 py-2 text-sm leading-5 rounded-full font-semibold text-white"
-          >
-            Answerforce US - WIP 🛠️
-          </Link>
-        </nav>
         <main>{children}</main>
       </body>
     </html>
