+++ src/app/page.tsx
@@ -1,6 +1,9 @@
+import Header from "./(components)/Header"
+
 export default function Home() {
   return (
     <main>
+      <Header />
       <div>
         <section>
          Email Signature Generator
