+++ src/app/signatures/[brand]/page.tsx
@@ -12,6 +12,7 @@ import ServiceForgeCaSignature from "@/app/SignatureHTML/ServiceForgeCaSignature
 import { InputsTypes } from "@/store/types";
 import { WRInputs } from "@/FormFields/WR";
 import { SFCAInputs } from "@/FormFields/SFCA";
+import Link from "next/link";
 
 const Brand = ({ params }: { params: { brand: string } }) => {
   const { data, setData } = useSignatureStore((state) => state);
@@ -43,11 +44,22 @@ const Brand = ({ params }: { params: { brand: string } }) => {
 
   return (
     <section>
-      <div className="container mx-auto py-10">
+      <div className="flex p-4 relative mb-4">
+        <Link
+          href="/"
+          className="inline-block px-4 py-2 bg-sky-500 text-slate-100 rounded-lg absolute top-2 left-2"
+        >
+          Back to homepage
+        </Link>
+        <h1 className="text-center flex-1 text-3xl">Email signature generator</h1>
+      </div>
+      <div className="container mx-auto pb-10">
         <div className="bg-gray-900 text-slate-100 p-8 rounded-lg w-full">
-          <h1 className="mb-8 text-2xl">
-            {params.brand.toUpperCase()} Signature Form
-          </h1>
+          <div>
+            <h2 className="mb-8 text-2xl">
+              {params.brand.toUpperCase()} Signature Form
+            </h2>
+          </div>
           <div className="flex justify-between gap-10 relative">
             <div className="w-1/2">
               <Form
